requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
click==7.1.1
discord.py==1.3.3
Flask==1.1.2
ftputil==3.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.5
pip==20.0.2
psutil==5.7.0
python-dotenv==0.13.0
setuptools==46.1.3
websockets==8.1
Werkzeug==1.0.1
wheel==0.34.2
yarl==1.4.2